Convert map to interactive leaflet choropleth
There's a lot of work that's gone into this commit. You can see with a git diff that the old functionality is still supported and commented out in the script (but it was pretty terrible and will likely be phased out in a future commit). The choropleth can't draw the polygons itself like the old map can - it only takes SpatialPolygons objects. The code strips away NA values (lat/lon coordinates belonging to a country we have not imported from) and creates Polygon objects for each country based on the map_data package's lat/lon coordinates. The work done for the old map to go from ISO Alpha 2 codes to full coordinate/value/country dataframes is still relevant