Add Export support for Non-EU Trade
This commit was all about getting the UI to support a selector between Import and Export trade data for the Non-EU databases. Some code needed to be refactored to support this - differentiating by 'source' and 'target' doesn't work when the sources and targets flip when you change between imports and exports. So, I refactored data frame labelling and data wrangling routines to work with unambiguous labels like 'country' and 'port', and pushed source/target differentiation to the definitions of the Link and Nodes dataframes (for the Sankey Diagram)