add .env database connection variables, add unit selection features
This is a major commit that's slightly incomplete - the functionality is now there to select between the different units, but it's buggy. Price per kilo is unbalanced, selecting the dateSlider element will crash the app because it triggers the reactive data shaping code, and exports are untested. In order to get this to work the Shape data had to be put into a different observeEvent, which is triggered every time a different unit/month is selected. All of the data is queried separately, at the click of the queryButton action button.

The reason this commit is being made is to add support for a .env file to store database credentials, which is now used in `InitialiseDataTables.R`, `InitialiseMetadataTables.R`, and `Shiny/app.R`. This will now be added to .gitignore to protect my personal credentials."