Add full unit selection and date filter features. Fix import/export country code column.
Following the buggy unit selection and date filtering features added to the code in the last commit, this commit has fixes for these features, so they work fully. It also contains a bit of code to select the correct country column for import or export. For imports, we want Country of Origin. For exports, Country of Origin doesn't exist, that wouldn't make sense. So it uses Country of Dispatch (ie, which country we're dispatching to). There's an additional bit of conditional code inserted to use the correct column in the group_by statement now.