Fix no-data crash
Previously when no data was returned from the database the entire app would crash since all the functions that shape the data were working on 0x0 dataframes. I've enabled checks for this (in the form of a req statement and new reactive variable nullDataframe) which mitigate this. When nothing is returned a modalDialog is displayed telling the user there is no data for those selections. For some reason the modal dialog doesn't show after a successful query is returned, this will be investigated.