Add Country and Port Selectors
I have refactored the UI a bit to support the addition of country and port selectors in the query. This required using names, rather than codes (as FXT is not obviously Felixstowe, XS is not obviously Serbia, etc). Additionally I have been able to reproduce some shinyapps.io bugs locally, which will now be investigated.