Create MonthlyUpdate.R
This script allows you to insert a specific month's data into the database, fully automated from download, to data transformation and cleaning, to dataload. It additionally tests whether that month exists in the database by checking for 50 records with that particular month's timestamp; since imports/exports to both EU and Non-EU countries both total over 100000 on average per month, this is sufficiently small to not risk overwriting but still permits the odd typo here and there in previous months' data.

Lastly, there's a small modification to the InitialiseMetadataTables.R script that uses the CNCodes.csv in the working directory to build the hierarchy, as EUROSTAT has been flaky lately.