Improve Date Slider Functionality
This commit has three main changes:

Improved Data Slider functionality - now has a separate All checkbox and more reliable reactivity with the visualisations

Reorganised filter ui elements - removed the Weight and Price/Weight ratio filters for the EU Trade Tab as it is unreliable data driving them; all functionality remains in the code and can be removed at a late later date if necessary, but it's dangerous to have for now since the data is so dirty. Additionally, the All checkbox and Date slider are now in separate columns so they can be 'in line'.

Extra control logic around null data when filtering. So if you select a month and it doesn't have data, the modal dialog shows up correctly.