More robust database management connections
I have used poolCheckout and poolReturn to create greater control around
the database management in an attempt to help shore up the reliability
of the application. It seems to be working from a first perspective,
I've run a few queries after the other, but my experiences may not be
the same as others. Will pass on for QA soon.

Also, a Beta Testing sheet (written in Org) has been added to the QA
folder.