Full Date Handling revamp etc
This commit completely overhauls the way that dates are handled in the
application. The master format is now YYYY-MM, with the Query type for
Non-EU being MM/YYYY, and the Query type for EU being 0YYYYMM. We now
have correctly ordered dates in the selectors, and it's just a lot more
intuitive to use as a whole. It contains a 2.5 month time lag to ensure
only available dates can be populated.

Also, I've changed some stuff around in the Map and Comcode Selectors.
The Map now displays the legend on the top right hand corner, has
different centering/zooming for both EU and non-EU. Comcode Selectors
now use blank entry to be the same as "All" being selected.