Show correct Seaports
This commit contains a nifty extra bit of filtering right at the top
when the portcode table is imported that strips out the minor jetties
and whatnot surrounding a main seaport. After discussion with Giles, who
indicated that, for non-unique portcodes, the **capitalised** portcode
should be the one that is shown as the port in the visualisation. This
is to avoid issues like Rochford seemingly importing a whole load of
Extra Virgin Olive Oil, when in fact it was Southend.

Also, a little bit of extra logic showing Origin/Dispatch in the tooltip
for the Leaflet Map for Non-EU. This helps differentiate between
Origin/Dispatch when using Import/Export data respectively.