Refactor Non-EU Query, Add Download Button
This commit contains a lot of code changes, because it involves an
overhaul of the way that Non-EU data is queried from the database. It
now queries port, comcode and country at the same time, and performs the
necessary GROUP_BY statements etc in R itself. This single query helps
cut down on query time by around half, as the returned dataframes are
rarely large enough to be computationally expensive compared to the time
it takes to obtain two queries from a remote database.

By doing things this way, we end up with both Non-EU and EU dataframes
close to what we'd want in a download. So, when the new "download"
buttons are hit (found in the bottom right hand corner of the screen),
you can simply obtain descriptions from the portcode, comcode and
countrycode dataframes, join them into the dataframe, and output to .csv
format. This was a major feature request, and is now implemented and
tested.