Recursively Search for Commodity Codes
A longstanding issue now solved; how to get the commodity codes to
recursively search for children.

The main changes in the commit is the addition of a `descendants(data,
code)` function, adapted from a "find all children" script written by
Arthur. This is a recursive algorithm, so it only has to be called once
(keeping code clean and ensuring optimal run time).

Using this function, I've completely refactored the comcode cascade
logic (for both Non-Eu and Eu, obviously). It's a lot cleaner, and
populates all *descendants*, not just *children*.

Performance was much better than expected - a query for all descendants
of commodity codes "01" and "02" took 0.25 seconds. Using server = TRUE
introduced bugs (flickering, poor performance), so I've left that
"improvement" out.