Add Feedback Tab, and small fixes
I've taken my Beta Testing form from QASpec/BetaTesting.org and adapted
it to a webform using WuFoo. I've embedded this in a new tab called
"Feedback", so users won't ever have to leave the app to provide
feedback.

I've also fixed an issue with the download button clipping itself off
to the right hand side of the screen, and added a warning on the landing
page saying to only use Chrome and that it's currently in Beta Testing.
Any bugs... go to the feedback tab!