Fix issue with certain countries missing from map
This was due to the iso.expand() formula for countries with overseas
territories returning a regexp with all of those territories and their
sovereign country. This can't be contracted with iso.alpha for matching,
so it was returning huge numbers with NAs (spanish olive oil, for
instance, because they have the Canaries). This has been fixed by
refactoring the mapWorld logic to circumvent this issue.