Add welcome page, fiddle with db connections
Pretty simple commit really, but with a fair amount of content. I've
added a welcome "splash" page. This contains a mini-tutorial, an intro
to the app, and an About section with author information and contact.
Have also moved the Beta alert there. Need to think about making a
loader so it doesn't look like it's frozen.

I've also changed the dbPool settings in the hope that it will make the
application more reliable; it's better than the last iteration but still
not perfect, so I'm going to continue changing these settings until we
have something that works nicely.