Resolve DB Crashes, some splash page fixes
I've changed around the way pool works in the application to help
prevent the database from becoming overloaded when a query crashes.
There was an astonishing amount of leaked connections, leading to the
max being hit very quickly and then dying for 15+ minutes. This problem
shouldn't happen again... But then again, I've said that before, and it
reared its ugly head once more.

I've also added existing bugs to the splash page and done a couple of
other housekeeping related tasks. We're now good to go for a public beta
amongst FSA stakeholders.