Add Lat/Lon to ports table, remove all minor ports
This commit pushes the filtering out of HMRC's minor ports to
`InitialiseMetadataTables.R`, and finishes off the addition of
latitude/longitude fields to the port table. It should be noted that
seaports are fairly off-base, as UN/Locode only goes to 2 d.p. for
latitude and 3 d.p. for longitude. When mapping at the UK level, this
results in inconsistencies (*many* seaports can be found in the Channel,
all of a sudden...). A bit more work may be done to fix this, or maybe
not. We will see.