Tidy repo, Start Documentation
The repo was filled with lots of legacy scripts that had no use anymore,
so I've cleared them out.

I've now also started on overhauling the documentation. The state as of
this commit is 80% or so, and I'll be finishing it this week. I've also
converted to the `.org` format, since it's easier to write in than
markdown and can be converted easily back if necessary.