Create `CheckLoad.R` and `CheckFoodBalance.R`
These are two scripts, now placed in folder *Utilities*, which allow you
to check the status of the hmrc database through running queries on how
many records there are for certain months of data, etc. Fun for datavis,
too.

I've also corrected the Updater to work on the remote database, and
moved `MonthlyUpdate.R` to the `superceded/` folder.