Include eu/non_eu column definitions for country
Works by querying what's in EU and Non-EU tables. Note - due to lots
of "country of origin" countries on the imports side, there may be EU
imports in the Non-EU countries selector. This shouldn't be much of a
problem though; the main issue was removing unnecessary countries from
the EU tab.