Fix `DownloadData.R` tilde issue on Windows machines
Windows cries for no reason when a tilde is included in the filename
and replaces anything matching the tilde as a wildcard that's already
in the directory. How utterly nonsensical. Anyway, fixed by changing
the order of operations so the offending tilde filenames are
processed *first*. And ignoring them in the next iteration.