Include instructions to build custom NetworkD3 that works w/ Modals.
This is largely a symbolic commit - it was at this point that I
managed to get the notifications, modals and progress bars working
correctly once the Sankey Network was rendered for the first time -
thus fixing one of the major UI bugs that I've been having. I went in
and made the changes found in PR 215:
https://github.com/christophergandrud/networkD3/pull/215/files

Once this has been done, you can recompile the NetworkD3 package and
then it'll work correctly with Modals, Notifications and Progress
Bars. Success - longest standing bug squished!