Filter out EU countries from Non-EU Country Selection
Because it was confusing to have Germany and France etc in the Non-EU
category of the dropdown and in the EU category too - Now it's clearer
that it's sorted by their EU/Non-EU status, not necessarily where the
data is. However, UK is entered as Non-EU... I suppose that won't be a
problem eventually.